core: remind: (fix) handle reminders for more than 24 days in the future
We use `setTimeout` for schedule reminders, but this function internally
uses a 32b integer to represent the delay (in milliseconds, which means
the maximum delay is roughly 24 days. To work around this limitation,
this commit changes the `launchReminder` function to schedule timeouts
for a maximum of 2^31 - 1 ms, and changes the `remind` to take the
intended reminder date as an argument and decide whether or not to
actually send the reminder, or to silently re-launch a new timeout if it
was called to early.